requirements.txt
streamlit
langchain_google_genai
python-dotenv